pyproject.toml
[project]
name = "smart-email-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.9.2",
    "beautifulsoup4>=4.12.0",
    "html2text>=2020.1.16",
    "lxml>=4.9.0",
    "pyyaml>=6.0",
]
